remove code redundancy
